5288 Split out class QueryWebSocketMessageHandler
