+++ java/clients/src/main/java/sleeper/clients/query/QueryWebSocketClient.java
@@ -15,11 +15,6 @@
  */
 package sleeper.clients.query;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonSyntaxException;
 import org.java_websocket.client.WebSocketClient;
 import org.java_websocket.handshake.ServerHandshake;
 import org.slf4j.Logger;
@@ -32,36 +27,17 @@ import software.amazon.awssdk.http.auth.aws.signer.AwsV4HttpSigner;
 import software.amazon.awssdk.http.auth.spi.internal.signer.DefaultSignRequest;
 import software.amazon.awssdk.http.auth.spi.signer.SignedRequest;
 
-import sleeper.clients.query.exception.MessageMalformedException;
-import sleeper.clients.query.exception.MessageMissingFieldException;
-import sleeper.clients.query.exception.UnknownMessageTypeException;
-import sleeper.clients.query.exception.WebSocketClosedException;
-import sleeper.clients.query.exception.WebSocketErrorException;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
-import sleeper.core.schema.Schema;
 import sleeper.core.util.LoggedDuration;
 import sleeper.query.core.model.Query;
-import sleeper.query.core.model.QuerySerDe;
-import sleeper.query.runner.websocket.QueryWebSocketMessage;
-import sleeper.query.runner.websocket.QueryWebSocketMessageSerDe;
-import sleeper.query.runner.websocket.QueryWebSocketMessageType;
 
 import java.net.URI;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.TreeMap;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
-import java.util.stream.Collectors;
 
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.QUERY_WEBSOCKET_API_URL;
 import static sleeper.core.properties.instance.CommonProperty.REGION;
@@ -126,7 +102,7 @@ public class QueryWebSocketClient {
         return (instanceProperties, tableProperties) -> {
             String region = instanceProperties.get(REGION);
             URI serverUri = URI.create(instanceProperties.get(QUERY_WEBSOCKET_API_URL));
-            WebSocketMessageHandler messageHandler = new WebSocketMessageHandler(tableProperties.getSchema());
+            QueryWebSocketMessageHandler messageHandler = new QueryWebSocketMessageHandler(tableProperties.getSchema());
             LOGGER.info("Obtaining AWS IAM credentials...");
             AwsCredentials credentials = credentialsProvider.resolveCredentials();
             return new WebSocketQueryClient(region, serverUri, credentials, messageHandler);
@@ -141,10 +117,10 @@ public class QueryWebSocketClient {
         private final String region;
         private final URI serverUri;
         private final AwsCredentials credentials;
-        private final WebSocketMessageHandler messageHandler;
+        private final QueryWebSocketMessageHandler messageHandler;
         private Query query;
 
-        private WebSocketQueryClient(String region, URI serverUri, AwsCredentials credentials, WebSocketMessageHandler messageHandler) {
+        private WebSocketQueryClient(String region, URI serverUri, AwsCredentials credentials, QueryWebSocketMessageHandler messageHandler) {
             super(serverUri);
             this.region = region;
             this.serverUri = serverUri;
@@ -223,186 +199,5 @@ public class QueryWebSocketClient {
 
     public interface ClientCloser {
         void close() throws InterruptedException;
-    }
-
-    public static class WebSocketMessageHandler {
-        private final Gson serde = new GsonBuilder().create();
-        private final Set<String> outstandingQueries = new HashSet<>();
-        private final Map<String, List<String>> parentQueryIdToSubQueryIds = new HashMap<>();
-        private final Map<String, List<String>> subQueryIdToRows = new TreeMap<>();
-        private final QueryWebSocketMessageSerDe serDe;
-        private final QuerySerDe querySerDe;
-        private ClientCloser clientCloser;
-        private boolean queryComplete = false;
-        private boolean queryFailed = false;
-        private long totalRecordsReturned = 0L;
-        private CompletableFuture<List<String>> future;
-        private String currentQueryId;
-
-        public WebSocketMessageHandler(Schema schema) {
-            this.serDe = QueryWebSocketMessageSerDe.withNoBatchSize(schema);
-            this.querySerDe = new QuerySerDe(schema);
-        }
-
-        public void setCloser(ClientCloser clientCloser) {
-            this.clientCloser = clientCloser;
-        }
-
-        public void setFuture(CompletableFuture<List<String>> future) {
-            this.future = future;
-        }
-
-        public void onOpen(Query query, Consumer<String> messageSender) {
-            LOGGER.info("Connected to WebSocket API");
-            String queryJson = querySerDe.toJson(query);
-            LOGGER.info("Submitting Query: {}", queryJson);
-            messageSender.accept(queryJson);
-            outstandingQueries.add(query.getQueryId());
-            parentQueryIdToSubQueryIds.put(query.getQueryId(), new ArrayList<>());
-            currentQueryId = query.getQueryId();
-        }
-
-        public void onMessage(String json) {
-            QueryWebSocketMessage messageNew;
-            try {
-                messageNew = serDe.fromJson(json);
-            } catch (RuntimeException e) {
-                future.completeExceptionally(e);
-                close();
-                return;
-            }
-            Optional<JsonObject> messageOpt = deserialiseMessage(json);
-            if (!messageOpt.isPresent()) {
-                close();
-                return;
-            }
-            JsonObject message = messageOpt.get();
-            String messageType = message.get("message").getAsString();
-            String queryId = message.get("queryId").getAsString();
-
-            if (messageNew.getMessage() == QueryWebSocketMessageType.error) {
-                handleError(messageNew);
-            } else if (messageNew.getMessage() == QueryWebSocketMessageType.subqueries) {
-                handleSubqueries(messageNew);
-            } else if (messageNew.getMessage() == QueryWebSocketMessageType.rows) {
-                handleRows(message, queryId);
-            } else if (messageNew.getMessage() == QueryWebSocketMessageType.completed) {
-                handleCompleted(messageNew);
-            } else {
-                queryFailed = true;
-                future.completeExceptionally(new UnknownMessageTypeException(messageType));
-                close();
-            }
-
-            if (outstandingQueries.isEmpty()) {
-                queryComplete = true;
-                future.complete(getResults(currentQueryId));
-                close();
-            }
-        }
-
-        private Optional<JsonObject> deserialiseMessage(String json) {
-            try {
-                JsonObject message = serde.fromJson(json, JsonObject.class);
-                if (!message.has("queryId")) {
-                    queryFailed = true;
-                    future.completeExceptionally(new MessageMissingFieldException("queryId", json));
-                    return Optional.empty();
-                }
-                if (!message.has("message")) {
-                    queryFailed = true;
-                    future.completeExceptionally(new MessageMissingFieldException("message", json));
-                    return Optional.empty();
-                }
-                return Optional.of(message);
-            } catch (JsonSyntaxException e) {
-                queryFailed = true;
-                future.completeExceptionally(new MessageMalformedException(json));
-                return Optional.empty();
-            }
-        }
-
-        private void handleError(QueryWebSocketMessage message) {
-            outstandingQueries.remove(message.getQueryId());
-            queryFailed = true;
-            future.completeExceptionally(new WebSocketErrorException(message.getError()));
-            close();
-        }
-
-        private void handleSubqueries(QueryWebSocketMessage message) {
-            LOGGER.info("Query {} split into the following subQueries:", message.getQueryIds());
-            for (String subQueryId : message.getQueryIds()) {
-                LOGGER.info("  " + subQueryId);
-                outstandingQueries.add(subQueryId);
-            }
-            outstandingQueries.remove(message.getQueryId());
-            parentQueryIdToSubQueryIds.compute(message.getQueryId(), (query, subQueries) -> {
-                subQueries.addAll(message.getQueryIds());
-                return subQueries;
-            });
-        }
-
-        private void handleRows(JsonObject message, String queryId) {
-            JsonArray recordBatch = message.getAsJsonArray("rows");
-            List<String> recordList = recordBatch.asList().stream()
-                    .map(jsonElement -> jsonElement.getAsJsonObject())
-                    .map(JsonObject::toString)
-                    .collect(Collectors.toList());
-            if (!subQueryIdToRows.containsKey(queryId)) {
-                subQueryIdToRows.put(queryId, recordList);
-            } else {
-                subQueryIdToRows.get(queryId).addAll(recordList);
-            }
-        }
-
-        private void handleCompleted(QueryWebSocketMessage message) {
-            long returnedRowCount = subQueryIdToRows.getOrDefault(message.getQueryId(), List.of()).size();
-            if (message.isRowsReturnedToClient() && message.getRowCount() > 0) {
-                if (returnedRowCount != message.getRowCount()) {
-                    LOGGER.error("API said it had returned {} rows for query {}, but only received {}",
-                            message.getRowCount(), message.getQueryId(), returnedRowCount);
-                }
-            }
-            outstandingQueries.remove(message.getQueryId());
-            LOGGER.info("{} rows returned by query {}. Remaining pending queries: {}",
-                    message.getRowCount(), message.getQueryId(), outstandingQueries.size());
-            totalRecordsReturned += returnedRowCount;
-        }
-
-        public void onClose(String reason) {
-            LOGGER.info("Disconnected from WebSocket API: {}", reason);
-            queryComplete = true;
-            future.completeExceptionally(new WebSocketClosedException(reason));
-        }
-
-        public void onError(Exception error) {
-            queryFailed = true;
-            future.completeExceptionally(new WebSocketErrorException(error));
-            close();
-        }
-
-        public boolean hasQueryFinished() {
-            return queryComplete || queryFailed;
-        }
-
-        public long getTotalRecordsReturned() {
-            return totalRecordsReturned;
-        }
-
-        public List<String> getResults(String queryId) {
-            return parentQueryIdToSubQueryIds.getOrDefault(queryId, List.of()).stream()
-                    .flatMap(id -> subQueryIdToRows.getOrDefault(id, List.of()).stream())
-                    .toList();
-        }
-
-        private void close() {
-            LOGGER.info("Query finished, closing client");
-            try {
-                clientCloser.close();
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new RuntimeException(e);
-            }
-        }
     }
 }
